requirements.txt
# Core dependencies
google-genai>=0.3.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Additional dependencies that may be needed
requests>=2.31.0
aiofiles>=23.0.0